code/debias_news_pipeline/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
jieba3k==0.35.1
joblib==0.17.0
lxml==4.6.1
newspaper3k==0.2.8
nltk==3.5
Pillow==8.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==3.0.2
tqdm==4.51.0
urllib3==1.25.11
websocket-client==0.48.0
